extra thicc fix

+++ lab2/PnmLine.h
@@ -5,6 +5,7 @@
 #define CGRAPH_PNMLINE_H
 
 void
-drawLine(int x0, int y0, int x1, int y1, PnmFile &pnm, double thick = 1, double gamma = 2.2, unsigned int br = 255);
+drawLine(double x0, double y0, double x1, double y1, PnmFile &pnm, double thick = 1, double gamma = 2.2,
+         unsigned int br = 255);
 
 #endif //CGRAPH_PNMLINE_H
